Rename to match the content of the properties
